+++ packages/ozone/src/mod-service/index.ts
@@ -18,6 +18,9 @@ import {
   isModEventTakedown,
   isModEventEmail,
   isModEventTag,
+  isAccountEvent,
+  isIdentityEvent,
+  isRecordEvent,
   REVIEWESCALATED,
   REVIEWOPEN,
 } from '../lexicon/types/tools/ozone/moderation/defs'
@@ -384,6 +387,25 @@ export class ModerationService {
       if (event.content) {
         meta.content = event.content
       }
+    }
+
+    if (isAccountEvent(event)) {
+      meta.active = event.active
+      meta.timestamp = event.timestamp
+      if (event.status) meta.status = event.status
+    }
+
+    if (isIdentityEvent(event)) {
+      meta.timestamp = event.timestamp
+      if (event.handle) meta.handle = event.handle
+      if (event.pdsHost) meta.pdsHost = event.pdsHost
+      if (event.tombstone) meta.tombstone = event.tombstone
+    }
+
+    if (isRecordEvent(event)) {
+      meta.timestamp = event.timestamp
+      meta.op = event.op
+      if (event.cid) meta.cid = event.cid
     }
 
     if (isModEventTakedown(event) && event.acknowledgeAccountSubjects) {
@@ -758,6 +780,11 @@ export class ModerationService {
     reportedAfter,
     reportedBefore,
     includeMuted,
+    hostingDeletedBefore,
+    hostingDeletedAfter,
+    hostingUpdatedBefore,
+    hostingUpdatedAfter,
+    hostingStatuses,
     onlyMuted,
     ignoreSubjects,
     sortDirection,
@@ -780,6 +807,11 @@ export class ModerationService {
     reportedAfter?: string
     reportedBefore?: string
     includeMuted?: boolean
+    hostingDeletedBefore?: string
+    hostingDeletedAfter?: string
+    hostingUpdatedBefore?: string
+    hostingUpdatedAfter?: string
+    hostingStatuses?: string[]
     onlyMuted?: boolean
     subject?: string
     ignoreSubjects?: string[]
@@ -845,6 +877,26 @@ export class ModerationService {
 
     if (reviewedBefore) {
       builder = builder.where('lastReviewedAt', '<', reviewedBefore)
+    }
+
+    if (hostingUpdatedAfter) {
+      builder = builder.where('hostingUpdatedAt', '>', hostingUpdatedAfter)
+    }
+
+    if (hostingUpdatedBefore) {
+      builder = builder.where('hostingUpdatedAt', '<', hostingUpdatedBefore)
+    }
+
+    if (hostingDeletedAfter) {
+      builder = builder.where('hostingDeletedAt', '>', hostingDeletedAfter)
+    }
+
+    if (hostingDeletedBefore) {
+      builder = builder.where('hostingDeletedAt', '<', hostingDeletedBefore)
+    }
+
+    if (hostingStatuses?.length) {
+      builder = builder.where('hostingStatus', 'in', hostingStatuses)
     }
 
     if (reportedAfter) {