+++ packages/sync/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @atproto/sync
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [[`1982693e3`](https://github.com/bluesky-social/atproto/commit/1982693e3ea1fef4db76ac9aca3db8dc5ebf3fe0)]:
+  - @atproto/xrpc-server@0.7.2
+  - @atproto/identity@0.4.3
+  - @atproto/repo@0.5.4
+
 ## 0.1.4
 
 ### Patch Changes